Build CLI: Force exit if graceful shutdown fails
If jobs_tracker timeout, the app will get into undefined behavior and
possible deadlocks. Previously the app will freeze and it wouldn't be
possible to close the app expect killing it from the OS.

The current solution will spawn another OS thread in this case, which
will wait for 2 Seconds then force the program to shutdown, while
printing useful information to `stderr` to inform the users about the
current state.