Build CLI: Ensure notarize succeed by check output
Notarize can fail while sending an OK status, therefore we check its
output if it contains the line where the succeed of the command is
stated.
This succeed line is set in the configuration to avoid having a new
release on the build tool each time apple change their command output.