Fix used filter counter
When user imports the counter in the Chipmunk, we show the counter
which indicates how many times imported filter is used. When user
clicks on the counter then counter keeps increasing even if filter
is already applied.
This commits fix that by checking if filter is applied or not,
if filter is used then, counter will not be increamented.

Fixes: #2430