Native: Run UI thread outside of tokio runtime
Having the UI thread running on tokio runtime and misusing
block_in_place is unpredictable and could cause many issues on different
platforms.
This commit moves spawning the runtime to the services thread and
passing a handle to the runtime to the UI thread so it can bridge
between the two worlds when needed.