Run Search on separate task
Search request are CPU intensive and will block the async runtime
especially if run with parsing in the same time (For example in
big DLT files) causing the app to lag.

Also I've added block_in_place in places where we expecting the code to
block due to intense CPU processes.